requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.28
numpy==1.26.4 